# swagger.yaml
openapi: 3.0.0
info:
  title: API de TFGs y Usuarios
  version: 1.0.0
  description: API para gestionar TFGs y usuarios en la plataforma
servers:
  - url: /api
paths:
  /tfgs:
    get:
      summary: Obtener todos los TFGs
      operationId: getTFGs
      tags:
        - TFGs
      responses:
        "200":
          description: Lista de TFGs obtenida correctamente
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/TFG"
    post:
      summary: Crear un nuevo TFG - Esta petición inicializa el Link como undefined, requiere un patch para añadir el PDF
      operationId: createTFG
      tags:
        - TFGs
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: array
              items:
                $ref: "#/components/schemas/TFGCreateAndUpdate"
      responses:
        "201":
          description: "TFG creado correctamente"
  /tfgs/names:
    get:
      summary: Obtener los nombres e ids de los TFGs
      operationId: getTFGnames
      tags:
        - TFGs
      responses:
        "200":
          description: Nombres de TFGs obtenido correctamente
          content:
            application/json:
              schema:
                type: object
                properties:
                  _id:
                    type: string
                  tfgTitle:
                    type: string
  /tfgs/{id}:
    get:
      summary: Obtener un TFG por ID
      operationId: getTFG
      tags:
        - TFGs
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: TFG obtenido correctamente
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TFG"
    put:
      summary: Actualizar un TFG completamente
      operationId: putTFG
      tags:
        - TFGs
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TFG"
      responses:
        "200":
          description: TFG actualizado correctamente
    delete:
      summary: Eliminar un TFG
      operationId: deleteTFG
      tags:
        - TFGs
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: TFG eliminado correctamente
  /tfgs/pdf/{id}:
    patch:
      summary: Actualizar el PDF de un TFG
      operationId: patchPDF
      tags:
        - TFGs
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                file:
                  type: string
                  format: binary
                filename:
                  type: string
      responses:
        "200":
          description: PDF actualizado correctamente
  /tfgs/pages/{page_number}:
    get:
      summary: Obtener los siguientes 10 TFGs con filtros opcionales
      operationId: getNextTFGs
      tags:
        - TFGs
      parameters:
        - name: page_number
          in: path
          required: true
          schema:
            type: integer
            minimum: 1
          description: Número de la página a obtener.
        - name: year
          in: query
          required: false
          schema:
            type: string
          description: Año del TFG.
        - name: degree
          in: query
          required: false
          schema:
            type: string
          description: Titulación del TFG.
        - name: student
          in: query
          required: false
          schema:
            type: string
          description: Nombre del estudiante.
        - name: tfgTitle
          in: query
          required: false
          schema:
            type: string
          description: Título del TFG (búsqueda por coincidencia parcial).
        - name: keywords
          in: query
          required: false
          schema:
            type: string
          description: Palabras clave separadas por comas (ej. "IA,redes,seguridad").
        - name: advisor
          in: query
          required: false
          schema:
            type: string
          description: Nombre del tutor.
        - name: abstract
          in: query
          required: false
          schema:
            type: string
          description: Búsqueda en el resumen del TFG (coincidencia parcial).
      responses:
        "200":
          description: Lista de TFGs obtenida correctamente
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/TFG"
        "500":
          description: Error interno del servidor
  /tfgs/verify/{id}:
    patch:
      summary: Verificar un TFG
      operationId: patchVerify
      tags:
        - TFGs
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                verified:
                  type: boolean
                reason:
                  type: string
      responses:
        "200":
          description: TFG verificado correctamente
  /users:
    get:
      summary: Obtener todos los usuarios
      operationId: getUsers
      tags:
        - Usuarios
      parameters:
        - name: Authorization
          in: header
          required: true
          schema:
            type: string
          description: Token de autenticación
      responses:
        "200":
          description: Lista de usuarios obtenida correctamente
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/User"
  /auth/register:
    post:
      summary: Crear un nuevo usuario
      operationId: createUser
      tags:
        - Usuarios
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/User"
      responses:
        "200":
          description: Usuario creado correctamente
  /auth/login:
    post:
      summary: Iniciar sesión
      operationId: loginUser
      tags:
        - Usuarios
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                email:
                  type: string
                password:
                  type: string
      responses:
        "200":
          description: Usuario logueado correctamente
  /users/{id}:
    get:
      summary: Obtener un usuario por ID
      operationId: getUser
      tags:
        - Usuarios
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Usuario obtenido correctamente
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
  /years:
    get:
      summary: Obtener los año academicos
      operationId: getYears
      tags:
        - years
      responses:
        "200":
          description: Años obtenidos correctamente
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Years"
    post:
      summary: Crear un nuevo año academico
      operationId: createYear
      tags:
        - years
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                year:
                  type: string
      responses:
        "201":
          description: Año académico creado correctamente
  /years/{id}:
    delete:
      summary: Eliminar un año academico
      operationId: deleteYear
      tags:
        - years
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Año eliminado correctamente

  /degrees:
    get:
      summary: Obtener los cursos academicos
      operationId: getDegrees
      tags:
        - degrees
      responses:
        "200":
          description: Curso obtenidos correctamente
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Degrees"
    post:
      summary: Crear un nuevo curso academico
      operationId: createDegree
      tags:
        - degrees
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                degree:
                  type: string
      responses:
        "201":
          description: Curso académico creado correctamente
  /degrees/{id}:
    delete:
      summary: Eliminar un curso academico
      operationId: deleteDegree
      tags:
        - degrees
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Curso eliminado correctamente
  /advisors:
    get:
      summary: Obtener los tutores
      operationId: getAdvisors
      tags:
        - advisors
      responses:
        "200":
          description: Tutores obtenidos correctamente
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Advisors"
    post:
      summary: Crear un nuevo curso academico
      operationId: createDegree
      tags:
        - advisors
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                advisor:
                  type: string
      responses:
        "201":
          description: Tutor creado correctamente
  /advisors/{id}:
    delete:
      summary: Eliminar un tutor
      operationId: deleteAdvisor
      tags:
        - advisors
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Tutor eliminado correctamente

components:
  schemas:
    TFG:
      type: object
      properties:
        id:
          type: string
        Año:
          type: string
        Titulación de Grado:
          type: string
        Alumno:
          type: string
        Título TFG:
          type: string
        Keywords:
          type: array
          items:
            type: string
        Link:
          type: string
        Tutor:
          type: string
        Abstract/Resumen:
          type: string
    TFGCreateAndUpdate:
      type: object
      properties:
        Año:
          type: string
        Titulación de Grado:
          type: string
        Alumno:
          type: string
        Título TFG:
          type: string
        Keywords:
          type: array
          items:
            type: string
        Tutor:
          type: string
        Abstract/Resumen:
          type: string
    User:
      type: object
      properties:
        name:
          type: string
        email:
          type: string
        password:
          type: string
    Years:
      type: object
      properties:
        year:
          type: string
        _id:
          type: string
    Degrees:
      type: object
      properties:
        degree:
          type: string
        _id:
          type: string
    Advisors:
      type: object
      properties:
        advisor:
          type: string
        _id:
          type: string